Refactored localfs platform specific code
